requirements.txt
streamlit
requests
pandas
plotly
textblob
python-dotenv
google-api-python-client
langdetect